Don't break in plot.sh if we're missing input files.

+++ R/consensus-stats.R
@@ -1,30 +1,37 @@
 options(warn = -1)
 suppressPackageStartupMessages(library("ggplot2"))
 
-relaysDay <- read.csv("stats/consensus-stats-raw", stringsAsFactors = FALSE)
-to <- Sys.time()
-from <- seq(from = to, length = 2, by = "-3 days")[2]
-relaysDay <- subset(relaysDay, as.POSIXct(datetime, tz = "GMT") >= from)
-m <- melt(relaysDay[,c(1, 5, 2)], id = "datetime")
-ggplot(m, aes(x = as.POSIXct(datetime, tz = "GMT"), y = value,
-  colour = variable)) + geom_point() +
-  scale_x_datetime(name = "", limits = c(from, to)) +
-  scale_y_continuous(name = "") +
-  scale_colour_hue("", breaks = c("running", "exit"),
-  labels = c("All relays", "Exit relays")) +
-  opts(title = "Number of exit relays (past 72 hours)\n")
-ggsave(filename = "website/graphs/exit/exit-72h.png",
-  width = 8, height = 5, dpi = 72)
+if (file.exists("stats/consensus-stats-raw")) {
+  relaysDay <- read.csv("stats/consensus-stats-raw",
+    stringsAsFactors = FALSE)
+  to <- Sys.time()
+  from <- seq(from = to, length = 2, by = "-3 days")[2]
+  relaysDay <- subset(relaysDay, as.POSIXct(datetime, tz = "GMT") >= from)
+  if (length(relaysDay$datetime) > 0) {
+    m <- melt(relaysDay[,c(1, 5, 2)], id = "datetime")
+    ggplot(m, aes(x = as.POSIXct(datetime, tz = "GMT"), y = value,
+      colour = variable)) + geom_point() +
+      scale_x_datetime(name = "", limits = c(from, to)) +
+      scale_y_continuous(name = "") +
+      scale_colour_hue("", breaks = c("running", "exit"),
+      labels = c("All relays", "Exit relays")) +
+      opts(title = "Number of exit relays (past 72 hours)\n")
+    ggsave(filename = "website/graphs/exit/exit-72h.png",
+      width = 8, height = 5, dpi = 72)
+  }
+}
 
-consensuses <- read.csv("stats/consensus-stats", header = TRUE,
-    stringsAsFactors = FALSE);
-consensuses <- consensuses[1:length(consensuses$date)-1,]
-write.csv(data.frame(date = consensuses$date,
-  relays = consensuses$running, bridges = consensuses$brunning),
-  "website/csv/networksize.csv", quote = FALSE, row.names = FALSE)
-write.csv(data.frame(date = consensuses$date,
-  all = consensuses$running, exit = consensuses$exit),
-  "website/csv/exit.csv", quote = FALSE, row.names = FALSE)
+if (file.exists("stats/consensus-stats")) {
+  consensuses <- read.csv("stats/consensus-stats", header = TRUE,
+      stringsAsFactors = FALSE);
+  consensuses <- consensuses[1:length(consensuses$date)-1,]
+  write.csv(data.frame(date = consensuses$date,
+    relays = consensuses$running, bridges = consensuses$brunning),
+    "website/csv/networksize.csv", quote = FALSE, row.names = FALSE)
+  write.csv(data.frame(date = consensuses$date,
+    all = consensuses$running, exit = consensuses$exit),
+    "website/csv/exit.csv", quote = FALSE, row.names = FALSE)
+}
 
 plot_consensus <- function(directory, filename, title, limits, rows, breaks,
     labels) {
@@ -139,9 +146,11 @@ plot_current <- function(directory, filenamePart, titlePart, rows, breaks,
   plot_all(directory, filenamePart, titlePart, rows, breaks, labels)
 }
 
-plot_current("website/graphs/networksize/", "networksize",
-  "Number of relays and bridges", c(1, 5, 7),
-  c("running", "brunning"), c("Relays", "Bridges"))
-plot_current("website/graphs/exit/", "exit", "Number of exit relays",
-  c(1, 5, 2), c("running", "exit"), c("All relays", "Exit relays"))
+if (file.exists("stats/consensus-stats")) {
+  plot_current("website/graphs/networksize/", "networksize",
+    "Number of relays and bridges", c(1, 5, 7),
+    c("running", "brunning"), c("Relays", "Bridges"))
+  plot_current("website/graphs/exit/", "exit", "Number of exit relays",
+    c(1, 5, 2), c("running", "exit"), c("All relays", "Exit relays"))
+}
